% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastPermFunctions.R
\name{diffMean}
\alias{diffMean}
\title{Utility function for fastPerm and SAMC}
\usage{
diffMean(x, y)
}
\arguments{
\item{x}{First sample}

\item{y}{Second sample}
}
\description{
Difference in means test statistic used by fastPerm
}
\examples{
diffMean(x, y )
}
\keyword{statistic}
\keyword{test}

