% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastPermFunctions.R
\name{diffMedian}
\alias{diffMedian}
\title{Utility function for fastPerm and SAMC}
\usage{
diffMedian(x, y)
}
\arguments{
\item{x}{First sample}

\item{y}{Second sample}
}
\description{
Difference in medians test statistic used by fastPerm.
Note: In this version of fastPerm, median statistic may not be reliable.
}
\examples{
diffMedian(x, y)
}
\keyword{statistic}
\keyword{test}

