% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastperm-package.R
\docType{package}
\name{fastperm-package}
\alias{fastperm}
\alias{fastperm-package}
\title{fastperm: Fast, General Permutation Testing with Sensitivity Analysis}
\description{
Fast permutation tests for arbitrary test statistics and as-if-randomized
designs, with optional Rosenbaum-style sensitivity analysis over Gamma.
See \code{docs/memo.md} for the design document. Methodology not yet implemented.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bowers-illinois-edu/fastperm}
  \item Report bugs at \url{https://github.com/bowers-illinois-edu/fastperm/issues}
}

}
\author{
\strong{Maintainer}: Jake Bowers \email{jake@jakebowers.org}

Authors:
\itemize{
  \item Jake Bowers \email{jake@jakebowers.org}
}

}
\keyword{internal}
