% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastPermFunctions.R
\name{fastPerm}
\alias{fastPerm}
\title{Fast approximation of small permutation p-values}
\usage{
fastPerm(x, y, testStat = ratioMean, B = 1000, adjusted = FALSE)
}
\arguments{
\item{x}{First sample}

\item{y}{Second sample}

\item{testStat}{Test statistic, defaults to the ratio of the
means (ratioMean). Other choices are DiffMean, ratioMedian, and
diffMedian. In the current version, the median statistics may
not be reliable.}

\item{B}{Number of Monte Carlo iterations within each partition. Defaults to 1000.}
}
\description{
This function approximates the p-value (two-sided) for a two
sample permutation test by partitioning the permutation space.
This function is most useful for small p-values, e.g. p < 10^-5.
}
\examples{
x <- rexp(100, 4)
y <- rexp(100, 2)
mStopRatioMean(x, y)
fastPerm(x, y)

x <- rnorm(110, 0, 1)
y <- rnorm(110, 1, 1)
mStopDiffMean(x, y)
fastPerm(x, y, testStat = diffMean)
}
\keyword{fast}
\keyword{permtution}
\keyword{sample}
\keyword{test}
\keyword{two}

