% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastPermAsymptoticFunctions.R
\name{fastPermAsym}
\alias{fastPermAsym}
\title{Fast approximation of small permutation p-values via asymptotic results}
\usage{
fastPermAsym(x, y, testStat = ratioMean, plot = FALSE)
}
\arguments{
\item{x}{First sample}

\item{y}{Second sample}

\item{testStat}{Test statistic, defaults to the ratio of the
means (ratioMean). The difference in means (diffMean) is also available.}

\item{B}{Number of Monte Carlo iterations within each partition. Defaults to 1,000.}
}
\description{
This function approximates the p-value (two-sided) for a two
sample permutation test by partitioning the permutation space,
and using asymptotic results on the p-values within each partition.
This function is most useful for small p-values, e.g. p < 10^-6, and
large sample sizes.
}
\examples{
x <- rexp(100, 5)
y <- rexp(100, 2)
mStopRatioMean(x, y)
fastPerm(x, y)

x <- rnorm(110, 0, 1)
y <- rnorm(110, 1, 1)
mStopDiffMean(x, y)
fastPerm(x, y, testStat = diffMean)
}
\keyword{fast}
\keyword{permtution}
\keyword{sample}
\keyword{test}
\keyword{two}

