% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAMCfunctions.R
\name{gammaFun}
\alias{gammaFun}
\title{Utility function for SAMC function}
\usage{
gammaFun(b, b0)
}
\arguments{
\item{b}{Current iteration of the SAMC algorithim}

\item{bo}{Iteration at which the updates begin to deteriorate.}
}
\description{
This function is used by SAMC to deteroriate the influence of updates
}
\examples{
gammaFun(b, b0)
}
\keyword{SAMC}
\keyword{utility}

