% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeComplexity_finite.R
\name{mHatFun}
\alias{mHatFun}
\title{Calculates the solution: mHat = m s.t. V'(m) = 0,
where V' is the derivative of Var(W|x,y) wrt m}
\usage{
mHatFun(x, y)
}
\description{
Calculates the solution: mHat = m s.t. V'(m) = 0,
where V' is the derivative of Var(W|x,y) wrt m
}

