% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeComplexity_finite.R
\name{mStopDiffMean}
\alias{mStopDiffMean}
\title{Estimate partition at which fastPerm will stop for difference statistics}
\usage{
mStopDiffMean(x, y, B = 1000, plot = TRUE)
}
\arguments{
\item{x}{First sample}

\item{y}{Second sample}

\item{B}{Number of Monte Carlo iterations within each partition. Defaults to 1000.}
}
\description{
This function estimates the partition at which fastPerm will
will stop, based on asymptotic approximation.
}
\examples{
x <- rnorm(100, 1)
y <- rexp(100, 0)
mStopDiffMean(x, y)
}
\keyword{fast}
\keyword{partition}
\keyword{permtution}
\keyword{stopping}

