% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastPermFunctions.R
\name{Pi}
\alias{Pi}
\title{Exact pmf of partitions}
\usage{
Pi(nx, ny)
}
\arguments{
\item{nx}{Size of first sample}

\item{ny}{Size of second sample}
}
\description{
This function calculates the exact probability mass function
of the partitions.
Typically, this function is called from fastPerm.
}
\examples{
Pi(nx, ny)
}
\keyword{partition}
\keyword{pmf}

