% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastPermFunctions.R
\name{PiLgammaNxSmaller}
\alias{PiLgammaNxSmaller}
\title{Exact pmf of partitions with the log gamma function
Assuming nx <= ny}
\usage{
PiLgammaNxSmaller(nx, ny)
}
\arguments{
\item{nx}{Size of first sample}

\item{ny}{Size of second sample}
}
\description{
This function calculates the exact probability mass function
of the partitions using the lgamma function.
Typically, this function is called from fastPerm.
}
\examples{
PiLgamma(nx, ny)
}
\keyword{partition}
\keyword{pmf}

