% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastPermFunctions.R
\name{PiStirling}
\alias{PiStirling}
\title{Stirling approximation of the pmf of partitions}
\usage{
PiStirling(nx, ny)
}
\arguments{
\item{nx}{Size of first sample}

\item{ny}{Size of second sample}
}
\description{
This function calculates the Stirling approximation to the
Typically, this function is called from fastPerm.
}
\examples{
PiStirling(nx, ny)
}
\keyword{Stirling}
\keyword{approximation}
\keyword{partition}
\keyword{pmf}

