% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastPermAsymptoticFunctions.R
\name{print.fastPermAsym}
\alias{print.fastPermAsym}
\title{Print function for fastPerm}
\usage{
\method{print}{fastPermAsym}(fp)
}
\arguments{
\item{fp}{Output from the fastPermAsym function}
}
\description{
This function prints the results of fastPermAsym
}
\examples{
x <- rexp(100, 5)
y <- rexp(100, 2)
print(fastPermAsym(x, y, testStat = ratioMean))
}
\keyword{fastPerm}
\keyword{print}

