% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastPermFunctions.R
\name{ratioMean}
\alias{ratioMean}
\title{Utility function for fastPerm and SAMC}
\usage{
ratioMean(x, y)
}
\arguments{
\item{x}{First sample}

\item{y}{Second sample}
}
\description{
Ratio of means test statistic used by fastPerm
}
\examples{
ratioMean(x, y)
}
\keyword{statistic}
\keyword{test}

