% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastPermFunctions.R
\name{stir}
\alias{stir}
\title{Utility function for PiStirling}
\usage{
stir(n)
}
\arguments{
\item{n}{Integer}
}
\description{
This function is used by the stirling function
}
\examples{
stir(n)
}
\keyword{stirling}
\keyword{utility}

