% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastPermFunctions.R
\name{stirling}
\alias{stirling}
\title{Utility function for PiStirling}
\usage{
stirling(nx, ny, m)
}
\arguments{
\item{nx}{Size of first sample}

\item{m}{Partition}

\item{nx}{Size of second sample}
}
\description{
This function is used by the PiStirling function to
calculate the mass in partition m
}
\examples{
stirling(nx, ny, m)
}
\keyword{stirling}
\keyword{utility}

