% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAMCfunctions.R
\name{swapFun}
\alias{swapFun}
\title{Utility function for SAMC function}
\usage{
swapFun(x, y, L)
}
\arguments{
\item{x}{First sample}

\item{y}{Second sample}

\item{L}{Number of elements to swap}
}
\description{
This function is used by SAMC to produce the next random permutation
}
\examples{
swapFun(x, y, L)
}
\keyword{SAMC}
\keyword{utility}

