% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeComplexity_finite.R
\name{print.mStop}
\alias{print.mStop}
\title{Print function for mStop functions}
\usage{
\method{print}{mStop}(mStopObj)
}
\arguments{
\item{mStopObj}{Output from the mStop functions}
}
\description{
This function prints the expected stopping partition
}
\examples{
x <- rnorm(110, 1)
y <- rnorm(95, 0)
print(mStopDiffMean(x, y))
}
\keyword{mStop}
\keyword{print}

